refactor: further match ticker cleanup
