+++ lua/wikis/commons/Infobox/UnofficialWorldChampion.lua
@@ -13,7 +13,7 @@ local Info = Lua.import('Module:Info', {loadData = true})
 local Json = Lua.import('Module:Json')
 local Logic = Lua.import('Module:Logic')
 local Lpdb = Lua.import('Module:Lpdb')
-local MatchTicker = Lua.import('Module:MatchTicker')
+local MatchTicker = Lua.import('Module:MatchTicker/Controller')
 local Namespace = Lua.import('Module:Namespace')
 local String = Lua.import('Module:StringUtils')
 local Table = Lua.import('Module:Table')
@@ -216,7 +216,7 @@ function UnofficialWorldChampion:_createUpcomingMatches()
 
 	return Logic.tryCatch(
 		function()
-			local matchTicker = MatchTicker{
+			return MatchTicker{
 				team = currentChampion.template,
 				limit = 5,
 				upcoming = true,
@@ -224,7 +224,6 @@ function UnofficialWorldChampion:_createUpcomingMatches()
 				hideTournament = false,
 				entityStyle = true,
 			}
-			return matchTicker:query():create()
 		end,
 		function()
 			return nil
