+++ lua/wikis/dota2/Infobox/Person/Player/Custom.lua
@@ -140,7 +140,7 @@ function CustomPlayer:adjustLPDB(lpdbData, args, personType)
 	return lpdbData
 end
 
----@return Widget?
+---@return Renderable?
 function CustomPlayer:createBottomContent()
 	if not Namespace.isMain() then
 		return
