+++ lua/wikis/starcraft2/Infobox/Person/Player/Custom.lua
@@ -157,7 +157,7 @@ function CustomPlayer:_getActiveCasterYears()
 	return YearsActive.displayYears(years:toArray())
 end
 
----@return Html?
+---@return Renderable?
 function CustomPlayer:createBottomContent()
 	if self.shouldQueryData then
 		return MatchTicker.recent({
